Adjust filters to be strict and address coding style comments
