--- hudi-client/hudi-spark-client/src/main/scala/org/apache/hudi/HoodieSparkWriterUtils.scala
@@ -1,64 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.hudi
-
-import java.util.Properties
-
-import org.apache.hudi.common.config.TypedProperties
-import org.apache.hudi.keygen.constant.KeyGeneratorOptions
-import org.apache.hudi.keygen.factory.HoodieSparkKeyGeneratorFactory
-import org.apache.hudi.keygen.{BaseKeyGenerator, CustomAvroKeyGenerator, CustomKeyGenerator, KeyGenerator}
-
-import scala.collection.JavaConverters._
-
-/**
- * A utility class for Spark writer
- */
-object HoodieSparkWriterUtils {
-  /**
-   * @param properties config properties
-   * @return partition columns
-   */
-  def getPartitionColumns(properties: Properties): String = {
-    val props = new TypedProperties(properties)
-    val keyGenerator = HoodieSparkKeyGeneratorFactory.createKeyGenerator(props)
-    getPartitionColumnsFromKeyGenerator(keyGenerator, props)
-  }
-
-  /**
-   * @param keyGen key generator
-   * @return partition columns
-   */
-  def getPartitionColumnsFromKeyGenerator(keyGen: KeyGenerator, typedProperties: TypedProperties): String = {
-    keyGen match {
-      // For CustomKeyGenerator and CustomAvroKeyGenerator, the partition path filed format
-      // is: "field_name: field_type", we extract the field_name from the partition path field.
-      case c: BaseKeyGenerator
-        if c.isInstanceOf[CustomKeyGenerator] || c.isInstanceOf[CustomAvroKeyGenerator] =>
-        c.getPartitionPathFields.asScala.map(pathField =>
-          pathField.split(CustomAvroKeyGenerator.SPLIT_REGEX)
-            .headOption.getOrElse(s"Illegal partition path field format: '$pathField' for ${c.getClass.getSimpleName}"))
-          .mkString(",")
-
-      case b: BaseKeyGenerator => b.getPartitionPathFields.asScala.mkString(",")
-      case _ => typedProperties.getString(KeyGeneratorOptions.PARTITIONPATH_FIELD.key())
-    }
-  }
-}
